Include full link details in link webhook events
